fix help menu bug, restructure ui folder a bit
