+++ src/aws_stuff_doer/uilib/__init__.py
@@ -1,3 +1,3 @@
 """The __init__.py filr for the ui module."""
 
-from .uilib import TerminalUi  # type: ignore
+from .uilib import AwsStuffDoer   # type: ignore
