+++ src/aws_stuff_doer/uilib/uilib.py
@@ -1,33 +1,24 @@
-from textual.app import App, ComposeResult
-from textual.events import Show
-from textual.widgets import Header, Footer
+from textual.app import App
+from aws_stuff_doer.uilib.screens import MainScreen, HelpScreen
 
+class AwsStuffDoer(App): # type: ignore
+    """beginnings of aws_stuff_doer terminal ui..."""
 
-class TerminalUi(App):
-    """A Textual app to manage stopwatches."""
+    MODES = { # type: ignore
+        "default": "main",
+        "help": "help",
+    }
 
+    SCREENS = {"main": MainScreen, "help": HelpScreen}
     TITLE = "ASD"
     SUB_TITLE = " An AWS Utility to help manage your projects and sso sessions."
+
     BINDINGS = [
         ("q", "quit", "Quit"),
         ("d", "toggle_dark", "Toggle dark mode"),
-        ("?", "show_help", "Show help"),
+        ("m", "switch_mode('default')", "Main"),
+        ("h", "switch_mode('help')", "Help"),
     ]
-    help = """
-    ASD: An AWS Utility to help manage your projects and sso sessions.
-
-    Keybindings:
-
-        d: Toggle dark mode
-        q: Quit
-        ?: Show help
-
-    """
-
-    def compose(self) -> ComposeResult:
-        """Create child widgets for the app."""
-        yield Header()
-        yield Footer()
 
     def action_toggle_dark(self) -> None:
         """An action to toggle dark mode."""
@@ -38,7 +29,9 @@ class TerminalUi(App):
             self.dark = True
             self.refresh()
 
+    def on_mount(self) -> None:
+        self.switch_mode("default")
 
 if __name__ == "__main__":
-    app = TerminalUi()
+    app = AwsStuffDoer()
     app.run()
